requirements.txt
django==5.0.2
pillow==10.2.0
djangorestframework==3.14.0